Correción en la navegación de pantallas
Muestra la pantalla de login al iniciar la aplicación, seguido del
welcome_screen que se acutalizó para mostrar los botones que navegan a
las otras pantallas de la aplicación.
